Changes:
1. Removed iteration without checking active state from Multiquery
2. Changed memory overhead on stable component by using uint32_t instead of uint64_t
3. Removed TConstSizeArray couse it was not used anywhere

+++ decs/src/decs/Container.cpp
@@ -15,7 +15,7 @@ namespace decs
 
 	Container::Container(
 		uint64_t enititesChunkSize,
-		uint64_t stableComponentDefaultChunkSize,
+		uint32_t stableComponentDefaultChunkSize,
 		uint64_t m_EmptyEntitiesChunkSize
 	) :
 		m_HaveOwnEntityManager(true),
@@ -29,7 +29,7 @@ namespace decs
 
 	Container::Container(
 		EntityManager* entityManager,
-		uint64_t stableComponentDefaultChunkSize,
+		uint32_t stableComponentDefaultChunkSize,
 		uint64_t m_EmptyEntitiesChunkSize
 	) :
 		m_HaveOwnEntityManager(entityManager == nullptr),
